agents/cv/requirements-ci.txt
jinja2>=3.1
pyyaml>=6.0
playwright>=1.53
